Refactor for elf
